pyproject.toml
[tool.poetry]
name = "chat-telegram-bot-ai-llamaindex"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.3"
llama-index = "^0.10.51"
chromadb = "^0.5.3"
sentence-transformers = "^3.0.1"
structlog = "^24.2.0"
tiktoken = "^0.7.0"
openai = "^1.35.7"
llama-index-llms-ollama = "^0.1.5"
llama-index-llms-openai = "^0.1.24"
llama-index-vector-stores-chroma = "^0.1.10"
llama-index-embeddings-huggingface = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
